Receipt parser added.
You can upload a clear image of a receipt, and it will give you the total carbon footprint from light of the items purchased. But if you were to scan a receipt, then the scanned receipt needs to be of HD quality, otherwise it is not able to recognize the text properly. A major issue is that sometimes, or most of the times, it is unable to differentiate between keywords like "tip" or "delivery" and considers them as items also. It has a default value, a default carbon footprint value for it. I tried to correct it, but I was unable to. If anyone can, that would be great.